genai-services/src/vision_ai/requirements.txt
fastapi
uvicorn
openai
pillow
python-multipart
google-generativeai
python-dotenv
pyjwt